# phi3 : duplicate rope factors in each layer