| GPU         | Model         | Test   |   t/s master |   t/s sl/phi3-fix |   Speedup |
|:------------|:--------------|:-------|-------------:|------------------:|----------:|
| RTX 3090 Ti | phi3 14B Q8_0 | pp512  |      1655.80 |           2359.53 |      1.43 |
| RTX 3090 Ti | phi3 14B Q8_0 | tg128  |        16.97 |             53.37 |      3.14 |